Reorganize test directory layout
