+++ Towny/src/main/java/com/palmergames/bukkit/towny/object/AbstractRegistryList.java
@@ -11,7 +11,6 @@ import org.bukkit.NamespacedKey;
 import org.bukkit.Registry;
 import org.bukkit.Tag;
 import org.jetbrains.annotations.NotNull;
-import org.jetbrains.annotations.VisibleForTesting;
 import java.util.Collection;
 import java.util.HashSet;
 import java.util.Locale;
@@ -52,8 +51,7 @@ public abstract class AbstractRegistryList<T extends Keyed> {
 		return matched != null && this.contains(matched);
 	}
 	
-	@VisibleForTesting
-	public Collection<T> tagged() {
+	protected Collection<T> tagged() {
 		return this.tagged;
 	}
