+++ Towny/src/test/java/com/palmergames/util/TimeTests.java
@@ -1,6 +1,5 @@
-package com.palmergames.bukkit.towny.test;
+package com.palmergames.util;
 
-import com.palmergames.util.TimeTools;
 import org.junit.jupiter.api.Test;
 
 import java.util.concurrent.TimeUnit;
